addad a database in sqlite, so only new files will be downloaded

+++ utilities.py
@@ -1,7 +1,9 @@
-import time
+import time 
+import datetime
 import urllib.request
 import os #creates f.e. folders
 from pathlib import Path
+import databasehandling as dbh
 
 #Just print out thre dots with a delay for a cool show
 def threedots():
@@ -28,7 +30,9 @@ def downloader(file_links, token, chosendata, chosenregion):
     
     while True:
         try:
-            save_location = Path.home()/"Downloads" #make the home Download folder the standard one 
+            dbh.dbcreation()
+
+            save_location = Path.home()/"Downloads" #make the home Download folder the default one 
 
             print("\nWhere do you want to save the data (press enter for C:\\Users\\[username]\\Downloads) \nWrite path like D:\\[folder]\\ ,or else data will be saved in the same folder as the script :")
             userlocation = input()
@@ -39,6 +43,8 @@ def downloader(file_links, token, chosendata, chosenregion):
 
             foldername = str(f"{save_location}/{chosendata}/{chosenregion}")
             
+
+
             print("\nStarting download", end="")
             threedots()
             print("\n")
@@ -49,40 +55,39 @@ def downloader(file_links, token, chosendata, chosenregion):
                 "Authorization": f"Bearer {token}" ##out downloadtoken into the request header
             }
 
+   
+            for i,dat in enumerate(file_links, start = 1):
 
-            dat = file_links[0]
-            filename = dat.split("/")[-1]
+                filename = dat.split("/")[-1]
 
-            filepath = os.path.join(foldername, filename) #create the correct path for the intended save location
+                year = int(filename[-11:-7])
 
-            req = urllib.request.Request(dat, headers = headers)
+                dayofyear=int(filename[-7:-4])
 
-            open(filepath, "wb").write(urllib.request.urlopen(req).read())
+                tmpdate = datetime.date(year,1,1) + datetime.timedelta(days=dayofyear -1)
 
-            break
+                currdate = tmpdate.strftime("%d.%m.%Y")
+
+                if dbh.dbexistfile(filename):
+                    print(f"The file {filename} already exists and will be skipped\n")
+                    continue 
+
+                filepath = os.path.join(foldername, filename) #create the correct path for the intended save location
+
+                req = urllib.request.Request(dat, headers = headers)
+
+                print("Downloading file ", i, "of ", len(file_links))
+
+                open(filepath, "wb").write(urllib.request.urlopen(req).read())
+            
+                dbh.dbdatainsert(filename,chosendata,chosenregion,currdate)
+                
 
         except FileNotFoundError:
             print("\nPlease enter a valid save location\n")
-
-
-
-
-
     
-    '''    for i,dat in enumerate(file_links, start = 1):
 
-        filename = dat.split("/")[-1]
-
-        filepath = os.path.join(foldername, filename) #create the correct path for the intended save location
-
-        req = urllib.request.Request(dat, headers = headers)
-
-        print("Downloading file ", i, "of ", len(file_links))
-
-        open(filepath, "wb").write(urllib.request.urlopen(req).read())
-    '''
-
-    return
+        return
 
 
 #################################################################
